+++ lib/widgets/net_banking/app_bar_search.dart
@@ -116,6 +116,7 @@ class AppBarSearch {
       } else if (onSubmitted != null) onSubmitted("");
     }
     // Display the change in the app's title
+    // ignore: invalid_use_of_protected_member
     state.setState(() {});
   }