+++ CHANGELOG.md
@@ -1,1 +1,18 @@
+## 1.0.1+5
+Styling option added for all the views.
+
+## 1.0.1+4
+Using flutter master channel solved the issue of webview keyboard input
+
+## 1.0.1+3
+1. Redirection handled from the server redirect url
+2. UPI decline case added.
+
+## 1.0.1+2
+Button and text filed color set to primary color set in theme of main.dart.
+
+## 1.0.1+1
+Bug Fixes
+
 ## 1.0.0
+Support Added for iOS and Android.